+++ lib/generic_lib.php
@@ -40,7 +40,40 @@ function jom_immediate_redirect($domain, $path, $page, $params = '') {
     $new_location .= '/' . $page;                           // page
     if ( !empty($params) ) $new_location .= '?'.$params;    // parameters
 
-    header('Location: '.$new_location);
+    header('Location: '.$new_location);                     // redirect immediately
+    die();                                                  // won't execute: redirected in the line above
+}
+
+
+/*
+ * Function: jom_clear_session
+ *   Clear session and cookie data and finally destroy the session
+ *
+ * Returns:
+ *   false if session is not initialized, true if everythig is fine
+ */
+function jom_clear_session() {
+
+    // check whether the session is started or not
+    if ( !isset($_SESSION) ) return false;
+
+    // Unset session variables
+    $_SESSION = array();
+
+    // If it's desired to kill the session, also delete the session cookie.
+    // Note: This will destroy the session, and not just the session data!
+    if (ini_get("session.use_cookies")) {
+        $params = session_get_cookie_params();
+        setcookie(session_name(), '', time() - 42000,
+            $params["path"], $params["domain"],
+            $params["secure"], $params["httponly"]
+        );
+    }
+
+    // Finally, destroy the session
+    session_destroy();
+
+    return true;
 }